orchestron-community-api/requirements.txt
Django==2.0
pytz==2017.3
djangorestframework==3.7.7
Pillow==5.0.0
lxml==4.1.1
django-background-tasks==1.1.12
bs4
jira==1.0.10
pycrypto==2.6.1
django-cors-headers==2.1.0
djangorestframework-filters==0.10.2
pep8==1.7.1
pylint==1.8.2
minio==3.0.2
djangorestframework-jwt==1.11.0
mysqlclient==1.3.12
django-queryset-csv==1.0.0
django-extensions==2.0.0
requests-cache==0.4.13